requirements.txt
django-paypal
paypal-checkout-serversdk
django-rest-framework
pymysql
mysqlclient
pillow
django-jazzmin
django-crispy-forms